+++ NGP_SERVER/Release/NGP_TermProject.log
@@ -1,0 +1,14 @@
+﻿  ServerMain.cpp
+C:\Users\user\Desktop\NGP텀프로젝트\NGP_SERVER\ServerMain.cpp(111,7): warning C4101: 'buf' :참조되지 않은 지역 변수입니다.
+C:\Users\user\Desktop\NGP텀프로젝트\NGP_SERVER\ServerMain.cpp(211,69): warning C4244: '인수': 'double'에서 'float'(으)로 변환하면서 데이터가 손실될 수 있습니다.
+C:\Users\user\Desktop\NGP텀프로젝트\NGP_SERVER\ServerMain.cpp(215,69): warning C4244: '인수': 'double'에서 'float'(으)로 변환하면서 데이터가 손실될 수 있습니다.
+C:\Users\user\Desktop\NGP텀프로젝트\NGP_SERVER\ServerMain.cpp(219,69): warning C4244: '인수': 'double'에서 'float'(으)로 변환하면서 데이터가 손실될 수 있습니다.
+C:\Users\user\Desktop\NGP텀프로젝트\NGP_SERVER\ServerMain.cpp(223,69): warning C4244: '인수': 'double'에서 'float'(으)로 변환하면서 데이터가 손실될 수 있습니다.
+  Session.cpp
+  error.cpp
+  코드를 생성하고 있습니다.
+  26 of 63 functions (41.3%) were compiled, the rest were copied from previous compilation.
+    0 functions were new in current compilation
+    2 functions had inline decision re-evaluated but remain unchanged
+  코드를 생성했습니다.
+  NGP_TermProject.vcxproj -> C:\Users\user\Desktop\NGP텀프로젝트\NGP_SERVER\Release\NGP_TermProject.exe
