+++ NGP_SERVER/Release/NGP_TermProject.tlog/NGP_TermProject.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v143:VCToolArchitecture=Native32Bit:VCToolsVersion=14.34.31933:TargetPlatformVersion=10.0.22621.0:
+Release|Win32|C:\Users\user\Desktop\NGP텀프로젝트\NGP_SERVER\|
