+++ NGP_SERVER/Release/NGP_TermProject.vcxproj.FileListAbsolute.txt
@@ -1,0 +1,1 @@
+C:\Users\user\Desktop\NGP텀프로젝트\NGP_SERVER\Release\NGP_TermProject.exe
