+++ NGP_SERVER/error.h
@@ -1,0 +1,6 @@
+#pragma once
+
+void err_quit(const char* msg);
+void err_display(const char* msg);
+void err_display(int errcode);
+
