+++ NGP_SERVER/x64/Debug/NGP_TermProject.log
@@ -1,0 +1,10 @@
+﻿  error.cpp
+  ServerMain.cpp
+C:\Users\user\Desktop\NGP_project\NGP_TermProject\NGP_SERVER\ServerMain.cpp(111,7): warning C4101: 'buf' :참조되지 않은 지역 변수입니다.
+C:\Users\user\Desktop\NGP_project\NGP_TermProject\NGP_SERVER\ServerMain.cpp(211,69): warning C4244: '인수': 'double'에서 'float'(으)로 변환하면서 데이터가 손실될 수 있습니다.
+C:\Users\user\Desktop\NGP_project\NGP_TermProject\NGP_SERVER\ServerMain.cpp(215,69): warning C4244: '인수': 'double'에서 'float'(으)로 변환하면서 데이터가 손실될 수 있습니다.
+C:\Users\user\Desktop\NGP_project\NGP_TermProject\NGP_SERVER\ServerMain.cpp(219,69): warning C4244: '인수': 'double'에서 'float'(으)로 변환하면서 데이터가 손실될 수 있습니다.
+C:\Users\user\Desktop\NGP_project\NGP_TermProject\NGP_SERVER\ServerMain.cpp(223,69): warning C4244: '인수': 'double'에서 'float'(으)로 변환하면서 데이터가 손실될 수 있습니다.
+  Session.cpp
+  코드를 생성하고 있습니다...
+  NGP_TermProject.vcxproj -> C:\Users\user\Desktop\NGP_project\NGP_TermProject\NGP_SERVER\x64\Debug\NGP_TermProject.exe
