+++ NGP_SERVER/x64/Debug/NGP_TermProject.tlog/NGP_TermProject.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.34.31933:TargetPlatformVersion=10.0.22621.0:
+Debug|x64|C:\Users\user\Desktop\NGP_project\NGP_TermProject\NGP_SERVER\|
