+++ NGP_SERVER/x64/Debug/NGP_TermProject.vcxproj.FileListAbsolute.txt
@@ -1,0 +1,1 @@
+C:\Users\user\Desktop\NGP_project\NGP_TermProject\NGP_SERVER\x64\Debug\NGP_TermProject.exe
