--- NetGamePrgm/NGP_TermProject.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33530.505
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NGP_TermProject", "NGP_TermProject.vcxproj", "{6580ECB7-A31A-4B40-B300-C28BB39E339D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6580ECB7-A31A-4B40-B300-C28BB39E339D}.Debug|x64.ActiveCfg = Debug|x64
-		{6580ECB7-A31A-4B40-B300-C28BB39E339D}.Debug|x64.Build.0 = Debug|x64
-		{6580ECB7-A31A-4B40-B300-C28BB39E339D}.Debug|x86.ActiveCfg = Debug|Win32
-		{6580ECB7-A31A-4B40-B300-C28BB39E339D}.Debug|x86.Build.0 = Debug|Win32
-		{6580ECB7-A31A-4B40-B300-C28BB39E339D}.Release|x64.ActiveCfg = Release|x64
-		{6580ECB7-A31A-4B40-B300-C28BB39E339D}.Release|x64.Build.0 = Release|x64
-		{6580ECB7-A31A-4B40-B300-C28BB39E339D}.Release|x86.ActiveCfg = Release|Win32
-		{6580ECB7-A31A-4B40-B300-C28BB39E339D}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9720EB99-D6F4-4BAE-8DC9-857FE14DD325}
-	EndGlobalSection
-EndGlobal
