--- NetGamePrgm/error.h
@@ -1,6 +1,0 @@
-#pragma once
-
-void err_quit(const char* msg);
-void err_display(const char* msg);
-void err_display(int errcode);
-
